review comments pt1
