+++ ring/lifecycler.go
@@ -38,7 +38,7 @@ var (
 	}, []string{"name"})
 	shutdownDuration = promauto.NewHistogramVec(prometheus.HistogramOpts{
 		Name:    "cortex_shutdown_duration_seconds",
-		Help:    "Duration (in seconds) of cortex shutdown procedure (ie transfer or flush).",
+		Help:    "Duration (in seconds) of shutdown procedure (ie transfer or flush).",
 		Buckets: prometheus.ExponentialBuckets(10, 2, 8), // Biggest bucket is 10*2^(9-1) = 2560, or 42 mins.
 	}, []string{"op", "status", "name"})
 )
@@ -371,7 +371,7 @@ func (i *Lifecycler) ClaimTokensFor(ctx context.Context, ingesterID string) erro
 			ing := ringDesc.Ingesters[i.ID]
 			ing.Timestamp = time.Now().Unix()
 
-			// Tokens of the leaving ingester may have been generated by an older version of Cortex which
+			// Tokens of the leaving ingester may have been generated by an older version which
 			// doesn't guarantee sorted tokens, so we enforce sorting here.
 			sort.Sort(tokens)
 			ing.Tokens = tokens
