+++ ring/model.go
@@ -434,7 +434,7 @@ func (d *Desc) getTokensInfo() map[uint32]instanceInfo {
 func (d *Desc) GetTokens() []uint32 {
 	instances := make([][]uint32, 0, len(d.Ingesters))
 	for _, instance := range d.Ingesters {
-		// Tokens may not be sorted for an older version of Cortex which, so we enforce sorting here.
+		// Tokens may not be sorted for an older version which, so we enforce sorting here.
 		tokens := instance.Tokens
 		if !sort.IsSorted(Tokens(tokens)) {
 			sort.Sort(Tokens(tokens))
@@ -451,7 +451,7 @@ func (d *Desc) GetTokens() []uint32 {
 func (d *Desc) getTokensByZone() map[string][]uint32 {
 	zones := map[string][][]uint32{}
 	for _, instance := range d.Ingesters {
-		// Tokens may not be sorted for an older version of Cortex which, so we enforce sorting here.
+		// Tokens may not be sorted for an older version which, so we enforce sorting here.
 		tokens := instance.Tokens
 		if !sort.IsSorted(Tokens(tokens)) {
 			sort.Sort(Tokens(tokens))
