+++ ring/ring.proto
@@ -37,7 +37,7 @@ message InstanceDesc {
 	// When an instance is already registered in the ring with a value of 0 it's NOT safe to
 	// update the timestamp to "now" because it would break the contract, given the instance
 	// was already registered before "now". If unknown (0), it should be left as is, and the
-	// Cortex code will properly deal with that.
+	// code will properly deal with that.
 	int64 registered_timestamp = 8;
 }
