+++ ring/tokens.go
@@ -79,7 +79,7 @@ func LoadTokensFromFile(tokenFilePath string) (Tokens, error) {
 	var t Tokens
 	err = t.Unmarshal(b)
 
-	// Tokens may have been written to file by an older version of Cortex which
+	// Tokens may have been written to file by an older version which
 	// doesn't guarantee sorted tokens, so we enforce sorting here.
 	if !sort.IsSorted(t) {
 		sort.Sort(t)
